refactor(lifespan): change lifespan archive for best comprehension
